bizzauto_api/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
sqlalchemy
psycopg2-binary
google-cloud-vision
redis
rq
python-multipart
openai